+++ apps/frontend/src/App.vue
@@ -1,5 +1,5 @@
 <script>
-import PageCatalog from './pages/PageCatalog.vue'
+import PageCatalog from '@/pages/PageProducts.vue'
 
 export default {
   components: { PageCatalog },
