+++ apps/frontend/src/components/page-products/BrickProduct.vue
