+++ apps/frontend/src/components/page-products/InputRange.vue
