+++ apps/frontend/src/components/page-products/ItemAttrValue.vue
