+++ apps/frontend/src/components/page-products/ResetAttributes.vue
