+++ apps/frontend/src/components/page-products/RowProduct.vue
