+++ apps/frontend/src/components/page-products/SelectCurrentPage.vue
