+++ apps/frontend/src/components/page-products/SelectListType.vue
