+++ apps/frontend/src/components/page-products/SelectPageSize.vue
