+++ apps/frontend/src/components/page-products/SelectSearchQuery.vue
