+++ apps/frontend/src/components/page-products/ShowFiltratedCount.vue
