+++ apps/frontend/src/pages/PageProducts.vue
@@ -1,18 +1,18 @@
 <script>
-import WidgetHeader from '@/components/WidgetHeader.vue'
+import WidgetHeader from '@/components/page-products/WidgetHeader.vue'
 import UiCategoryNavigation from '@/ui/UiCategoryNavigation.vue'
-import VerboseFiltration from '@/components/VerboseFiltration.vue'
-import WidgetProducts from '@/components/WidgetProducts.vue'
-import SelectCurrentPage from '@/components/SelectCurrentPage.vue'
-import WatchedProducts from '@/components/WatchedProducts.vue'
+import VerboseFiltration from '@/components/page-products/VerboseFiltration.vue'
+import WidgetProducts from '@/components/page-products/WidgetProducts.vue'
+import SelectCurrentPage from '@/components/page-products/SelectCurrentPage.vue'
+import WatchedProducts from '@/components/page-products/WatchedProducts.vue'
 import UiLinks from '@/ui/UiLinks.vue'
 import UiTags from '@/ui/UiTags.vue'
 import UiCopyright from '@/ui/UiCopyright.vue'
 import UiTapbar from '@/ui/UiTapbar.vue'
 import UiLinetop from '@/ui/UiLinetop.vue'
-import SelectPriceRanges from '@/components/SelectPriceRanges.vue'
-import SelectAttributes from '@/components/SelectAttributes.vue'
-import OffCanvas from '@/components/OffCanvas.vue'
+import SelectPriceRanges from '@/components/page-products/SelectPriceRanges.vue'
+import SelectAttributes from '@/components/page-products/SelectAttributes.vue'
+import OffCanvas from '@/components/page-products/OffCanvas.vue'
 
 import searchProducts from '@/functions/searchProducts'
 import rangeProducts from '@/functions/rangeProducts'
